Delete most uses of MarkerOwner[][]
It is never shown to the user, and main/multi.c mishandled it in a way
that could corrupt memory if player #2 drops his second marker or
players numbered 3 or above dropped any markers.

Delete the offending usage in main/multi.c.
Delete the use in main/automap.c since it is useless, though not
dangerous.